backend/requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
cryptography==41.0.7
openai==1.12.0
